// src/react-app/lib/finance-types.ts
export interface Company {
  id: string;
  name: string;
  cnpj?: string;
  address?: string;
}